userdb.{c,h}: extract user/group database related functions to a separate module
Instead of honing everything into misc.c, extract the user/group related
functions into a separate file.
