+++ lib/gprocess.c
@@ -23,7 +23,7 @@
  */
 
 #include "gprocess.h"
-#include "misc.h"
+#include "userdb.h"
 #include "messages.h"
 #include "reloc.h"
