+++ tests/unit/test_resolve_pwgr.c
@@ -1,4 +1,4 @@
-#include "misc.h"
+#include "userdb.h"
 
 int main(void)
 {
